Clean up JsonReader tests and split types into separate files
