+++ tests/System.Text.JsonLab.Tests/JsonTestHelper.cs
@@ -2,11 +2,16 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Buffers;
+using System.Buffers.Text;
+using System.Collections.Generic;
 using System.IO;
+using System.Runtime.InteropServices;
+using System.Text.Formatting;
 
 namespace System.Text.JsonLab.Tests
 {
-    public static class JsonTestHelper
+    internal static class JsonTestHelper
     {
         public static string NewtonsoftReturnStringHelper(TextReader reader)
         {
@@ -20,6 +25,688 @@ namespace System.Text.JsonLab.Tests
                 }
             }
             return sb.ToString();
+        }
+
+        public static void WriteDepth(ref Utf8JsonWriter<ArrayFormatterWrapper> jsonUtf8, int depth)
+        {
+            jsonUtf8.WriteObjectStart();
+            for (int i = 0; i < depth; i++)
+            {
+                jsonUtf8.WriteObjectStart("message" + i);
+            }
+            jsonUtf8.WriteAttribute("message" + depth, "Hello, World!");
+            for (int i = 0; i < depth; i++)
+            {
+                jsonUtf8.WriteObjectEnd();
+            }
+            jsonUtf8.WriteObjectEnd();
+            jsonUtf8.Flush();
+        }
+
+        public static string ObjectToString(object jsonValues)
+        {
+            string s = "";
+            if (jsonValues is List<object> jsonList)
+            {
+                s = ListToString(jsonList);
+            }
+            else
+            {
+                if (jsonValues is Dictionary<string, object> jsonDictionary)
+                {
+
+                    s = DictionaryToString(jsonDictionary);
+                }
+            }
+            return s;
+        }
+
+        public static string DictionaryToString(Dictionary<string, object> dictionary)
+        {
+            var builder = new StringBuilder();
+            foreach (KeyValuePair<string, object> entry in dictionary)
+            {
+                if (entry.Value is Dictionary<string, object> nestedDictionary)
+                {
+                    builder.Append(entry.Key).Append(", ").Append(DictionaryToString(nestedDictionary));
+                }
+                else if (entry.Value is List<object> nestedList)
+                {
+                    builder.Append(entry.Key).Append(", ").Append(ListToString(nestedList));
+                }
+                else
+                {
+                    builder.Append(entry.Key).Append(", ").Append(entry.Value).Append(", ");
+                }
+            }
+            return builder.ToString();
+        }
+
+        public static string ListToString(List<object> list)
+        {
+            var builder = new StringBuilder();
+            foreach (object entry in list)
+            {
+                if (entry is Dictionary<string, object> nestedDictionary)
+                {
+                    builder.Append(DictionaryToString(nestedDictionary));
+                }
+                else if (entry is List<object> nestedList)
+                {
+                    builder.Append(ListToString(nestedList));
+                }
+                else
+                {
+                    builder.Append(entry).Append(", ");
+                }
+            }
+            return builder.ToString();
+        }
+
+        public static void SetKeyValues(ref Utf8JsonReader json, Dictionary<string, object> dictionary, ref string key, ref object value)
+        {
+            while (json.Read())
+            {
+                switch (json.TokenType)
+                {
+                    case JsonTokenType.PropertyName:
+                        key = Encoding.UTF8.GetString(json.Value);
+                        dictionary.Add(key, null);
+                        break;
+                    case JsonTokenType.Value:
+                        switch (json.ValueType)
+                        {
+                            case JsonValueType.String:
+                                value = Encoding.UTF8.GetString(json.Value);
+                                break;
+                            case JsonValueType.False:
+                                value = false;
+                                break;
+                        }
+                        if (dictionary.TryGetValue(key, out _))
+                        {
+                            dictionary[key] = value;
+                        }
+                        else
+                        {
+                            dictionary.Add(key, value);
+                        }
+                        break;
+                    default:
+                        break;
+                }
+            }
+        }
+
+        public static bool TryParseResponseMessage(ref ReadOnlySequence<byte> buffer)
+        {
+            if (!TryParseMessage(ref buffer, out var payload))
+            {
+                return false;
+            }
+
+            var reader = new Utf8JsonReader(payload);
+
+            CheckRead(ref reader);
+            EnsureObjectStart(ref reader);
+
+            int? minorVersion = null;
+            string error = null;
+
+            var completed = false;
+            while (!completed && CheckRead(ref reader))
+            {
+                switch (reader.TokenType)
+                {
+                    case JsonTokenType.PropertyName:
+                        ReadOnlySpan<byte> memberName = reader.Value;
+
+                        if (memberName.SequenceEqual(TypePropertyNameUtf8))
+                        {
+
+                            // a handshake response does not have a type
+                            // check the incoming message was not any other type of message
+                            throw new InvalidDataException("Handshake response should not have a 'type' value.");
+                        }
+                        else if (memberName.SequenceEqual(ErrorPropertyNameUtf8))
+                        {
+                            error = ReadAsString(ref reader, ErrorPropertyName);
+                        }
+                        else if (memberName.SequenceEqual(MinorVersionPropertyNameUtf8))
+                        {
+                            minorVersion = ReadAsInt32(ref reader, MinorVersionPropertyName);
+                        }
+                        else
+                        {
+                            reader.Skip();
+                        }
+                        break;
+                    case JsonTokenType.EndObject:
+                        completed = true;
+                        break;
+                    default:
+                        throw new InvalidDataException($"Unexpected token '{reader.TokenType}' when reading handshake response JSON.");
+                }
+            };
+
+            return true;
+        }
+
+        public static readonly byte RecordSeparator = 0x1e;
+
+        public static bool TryParseMessage(ref ReadOnlySequence<byte> buffer, out ReadOnlySequence<byte> payload)
+        {
+            var position = buffer.PositionOf(RecordSeparator);
+            if (position == null)
+            {
+                payload = default;
+                return false;
+            }
+
+            payload = buffer.Slice(0, position.Value);
+
+            // Skip record separator
+            buffer = buffer.Slice(buffer.GetPosition(1, position.Value));
+
+            return true;
+        }
+
+        public static int? ReadAsInt32(ref Utf8JsonReader reader, string propertyName)
+        {
+            reader.Read();
+
+            if (reader.TokenType != JsonTokenType.Value || reader.ValueType != JsonValueType.Number)
+            {
+                throw new InvalidDataException($"Expected '{propertyName}' to be of type Integer.");
+            }
+
+            if (reader.Value.IsEmpty)
+            {
+                return null;
+            }
+            if (!Utf8Parser.TryParse(reader.Value, out int value, out _))
+            {
+                throw new InvalidDataException($"Expected '{propertyName}' to be of type Integer.");
+            }
+            return value;
+        }
+
+        public static bool CheckRead(ref Utf8JsonReader reader)
+        {
+            if (!reader.Read())
+            {
+                throw new InvalidDataException("Unexpected end when reading JSON.");
+            }
+
+            return true;
+        }
+
+        public static string GetTokenString(JsonValueType valueType, JsonTokenType tokenType)
+        {
+            switch (valueType)
+            {
+                case JsonValueType.Number:
+                    return "Integer";
+                case JsonValueType.Unknown:
+                    if (tokenType == JsonTokenType.StartArray)
+                    {
+                        return JsonValueType.Array.ToString();
+                    }
+                    if (tokenType == JsonTokenType.StartObject)
+                    {
+                        return JsonValueType.Object.ToString();
+                    }
+                    return tokenType.ToString();
+                default:
+                    break;
+            }
+            return valueType.ToString();
+        }
+
+        public static void EnsureObjectStart(ref Utf8JsonReader reader)
+        {
+            if (reader.TokenType != JsonTokenType.StartObject)
+            {
+                throw new InvalidDataException($"Unexpected JSON Token Type '{GetTokenString(reader.ValueType, reader.TokenType)}'. Expected a JSON Object.");
+            }
+        }
+
+        private const string ProtocolPropertyName = "protocol";
+        private const string ProtocolVersionPropertyName = "version";
+        private const string MinorVersionPropertyName = "minorVersion";
+        private const string ErrorPropertyName = "error";
+        private const string TypePropertyName = "type";
+
+        public static readonly byte[] ProtocolPropertyNameUtf8 = Encoding.UTF8.GetBytes("protocol");
+        public static readonly byte[] ProtocolVersionPropertyNameUtf8 = Encoding.UTF8.GetBytes("version");
+        public static readonly byte[] MinorVersionPropertyNameUtf8 = Encoding.UTF8.GetBytes("minorVersion");
+        public static readonly byte[] ErrorPropertyNameUtf8 = Encoding.UTF8.GetBytes("error");
+        public static readonly byte[] TypePropertyNameUtf8 = Encoding.UTF8.GetBytes("type");
+
+        public static bool TryParseRequestMessage(ref ReadOnlySequence<byte> buffer)
+        {
+            if (!TryParseMessage(ref buffer, out var payload))
+            {
+                return false;
+            }
+
+            var reader = new Utf8JsonReader(payload);
+            CheckRead(ref reader);
+            EnsureObjectStart(ref reader);
+
+            string protocol = null;
+            int? protocolVersion = null;
+
+            var completed = false;
+            while (!completed && CheckRead(ref reader))
+            {
+                switch (reader.TokenType)
+                {
+                    case JsonTokenType.PropertyName:
+                        ReadOnlySpan<byte> memberName = reader.Value;
+
+                        if (memberName.SequenceEqual(ProtocolPropertyNameUtf8))
+                        {
+                            protocol = ReadAsString(ref reader, ProtocolPropertyName);
+                        }
+                        else if (memberName.SequenceEqual(ProtocolVersionPropertyNameUtf8))
+                        {
+                            protocolVersion = ReadAsInt32(ref reader, ProtocolVersionPropertyName);
+                        }
+                        else
+                        {
+                            reader.Skip();
+                        }
+                        break;
+                    case JsonTokenType.EndObject:
+                        completed = true;
+                        break;
+                    default:
+                        throw new InvalidDataException($"Unexpected token '{reader.TokenType}' when reading handshake request JSON.");
+                }
+            }
+
+            if (protocol == null)
+            {
+                throw new InvalidDataException($"Missing required property '{ProtocolPropertyName}'.");
+            }
+            if (protocolVersion == null)
+            {
+                throw new InvalidDataException($"Missing required property '{ProtocolVersionPropertyName}'.");
+            }
+
+            return true;
+        }
+
+        public static unsafe string ReadAsString(ref Utf8JsonReader reader, string propertyName)
+        {
+            reader.Read();
+
+            if (reader.TokenType != JsonTokenType.Value || reader.ValueType != JsonValueType.String)
+            {
+                throw new InvalidDataException($"Expected '{propertyName}' to be of type String.");
+            }
+
+            if (reader.Value.IsEmpty) return "";
+
+#if NETCOREAPP2_2
+            return Encoding.UTF8.GetString(reader.Value);
+#else
+            fixed (byte* bytes = &MemoryMarshal.GetReference(reader.Value))
+            {
+                return Encoding.UTF8.GetString(bytes, reader.Value.Length);
+            }
+#endif
+        }
+
+        public static void JsonLabEmptyLoopHelper(byte[] data)
+        {
+            var json = new Utf8JsonReader(data);
+            while (json.Read())
+            {
+                JsonTokenType tokenType = json.TokenType;
+                switch (tokenType)
+                {
+                    case JsonTokenType.StartObject:
+                    case JsonTokenType.EndObject:
+                    case JsonTokenType.StartArray:
+                    case JsonTokenType.EndArray:
+                        break;
+                    case JsonTokenType.PropertyName:
+                        break;
+                    case JsonTokenType.Value:
+                        JsonValueType valueType = json.ValueType;
+                        switch (valueType)
+                        {
+                            case JsonValueType.Unknown:
+                                break;
+                            case JsonValueType.Object:
+                                break;
+                            case JsonValueType.Array:
+                                break;
+                            case JsonValueType.Number:
+                                break;
+                            case JsonValueType.String:
+                                break;
+                            case JsonValueType.True:
+                                break;
+                            case JsonValueType.False:
+                                break;
+                            case JsonValueType.Null:
+                                break;
+                        }
+                        break;
+                    case JsonTokenType.None:
+                        break;
+                    case JsonTokenType.Comment:
+                        break;
+                    default:
+                        throw new ArgumentException();
+                }
+            }
+        }
+
+        public static ReadOnlySequence<byte> CreateSegments(byte[] data)
+        {
+            ReadOnlyMemory<byte> dataMemory = data;
+
+            var firstSegment = new BufferSegment<byte>(dataMemory.Slice(0, data.Length / 2));
+            ReadOnlyMemory<byte> secondMem = dataMemory.Slice(data.Length / 2);
+            BufferSegment<byte> secondSegment = firstSegment.Append(secondMem);
+
+            return new ReadOnlySequence<byte>(firstSegment, 0, secondSegment, secondMem.Length);
+        }
+
+        public static byte[] JsonLabSequenceReturnBytesHelper(byte[] data, out int length)
+        {
+            ReadOnlySequence<byte> sequence = CreateSegments(data);
+            var reader = new Utf8JsonReader(sequence);
+            byte[] result = JsonLabReaderLoop(data.Length, out length, ref reader);
+            // TODO: Should we reset the value and valuetype once we are done?
+            //Assert.True(reader.Value.IsEmpty);
+            //Assert.Equal(JsonValueType.Unknown, reader.ValueType);
+            return result;
+        }
+
+        public static byte[] JsonLabReaderLoop(int inpuDataLength, out int length, ref Utf8JsonReader json)
+        {
+            byte[] outputArray = new byte[inpuDataLength];
+            Span<byte> destination = outputArray;
+
+            while (json.Read())
+            {
+                JsonTokenType tokenType = json.TokenType;
+                ReadOnlySpan<byte> valueSpan = json.Value;
+                switch (tokenType)
+                {
+                    case JsonTokenType.PropertyName:
+                        valueSpan.CopyTo(destination);
+                        destination[valueSpan.Length] = (byte)',';
+                        destination[valueSpan.Length + 1] = (byte)' ';
+                        destination = destination.Slice(valueSpan.Length + 2);
+                        break;
+                    case JsonTokenType.Value:
+                        JsonValueType valueType = json.ValueType;
+
+                        switch (valueType)
+                        {
+                            // Special casing True/False so that the casing matches with Json.NET
+                            case JsonValueType.True:
+                                destination[0] = (byte)'T';
+                                destination[1] = (byte)'r';
+                                destination[2] = (byte)'u';
+                                destination[3] = (byte)'e';
+                                destination[valueSpan.Length] = (byte)',';
+                                destination[valueSpan.Length + 1] = (byte)' ';
+                                destination = destination.Slice(valueSpan.Length + 2);
+                                break;
+                            case JsonValueType.False:
+                                destination[0] = (byte)'F';
+                                destination[1] = (byte)'a';
+                                destination[2] = (byte)'l';
+                                destination[3] = (byte)'s';
+                                destination[4] = (byte)'e';
+                                destination[valueSpan.Length] = (byte)',';
+                                destination[valueSpan.Length + 1] = (byte)' ';
+                                destination = destination.Slice(valueSpan.Length + 2);
+                                break;
+                            case JsonValueType.Number:
+                            case JsonValueType.String:
+                                valueSpan.CopyTo(destination);
+                                destination[valueSpan.Length] = (byte)',';
+                                destination[valueSpan.Length + 1] = (byte)' ';
+                                destination = destination.Slice(valueSpan.Length + 2);
+                                break;
+                            case JsonValueType.Null:
+                                // Special casing Null so that it matches what JSON.NET does
+                                break;
+                        }
+                        break;
+                    default:
+                        break;
+                }
+            }
+            length = outputArray.Length - destination.Length;
+            return outputArray;
+        }
+
+        public static object JsonLabReaderLoop(ref Utf8JsonReader json)
+        {
+            object root = null;
+
+            while (json.Read())
+            {
+                JsonTokenType tokenType = json.TokenType;
+                ReadOnlySpan<byte> valueSpan = json.Value;
+                switch (tokenType)
+                {
+                    case JsonTokenType.Value:
+                        JsonValueType valueType = json.ValueType;
+
+                        switch (valueType)
+                        {
+                            case JsonValueType.True:
+                            case JsonValueType.False:
+                                root = valueSpan.ConvertToBool();
+                                break;
+                            case JsonValueType.Number:
+
+                                NumberType type;
+                                (type, root) = valueSpan.GetNumberType();
+
+                                //OR
+
+                                /*if (Utf8Parser.TryParse(valueSpan, out int intVal, out int bytesConsumed))
+                                {
+                                    if (valueSpan.Length == bytesConsumed)
+                                    {
+                                        value = intVal;
+                                        break;
+                                    }
+                                }
+                                if (Utf8Parser.TryParse(valueSpan, out long longVal, out bytesConsumed))
+                                {
+                                    if (valueSpan.Length == bytesConsumed)
+                                    {
+                                        value = longVal;
+                                        break;
+                                    }
+                                }
+                                if (valueSpan.IndexOfAny((byte)'.', (byte)'e', (byte)'E') != -1)
+                                {
+                                    value = valueSpan.ConvertToDecimal();
+                                }*/
+                                break;
+                            case JsonValueType.String:
+                                root = valueSpan.ConvertToString();
+                                break;
+                            case JsonValueType.Null:
+                                break;
+                        }
+
+                        break;
+                    case JsonTokenType.StartObject:
+                        root = JsonLabReaderDictionaryLoop(ref json);
+                        break;
+                    case JsonTokenType.StartArray:
+                        root = JsonLabReaderListLoop(ref json);
+                        break;
+                    case JsonTokenType.EndObject:
+                    case JsonTokenType.EndArray:
+                        break;
+                    case JsonTokenType.None:
+                    case JsonTokenType.Comment:
+                    default:
+                        break;
+                }
+            }
+            return root;
+        }
+
+        public static Dictionary<string, object> JsonLabReaderDictionaryLoop(ref Utf8JsonReader json)
+        {
+            Dictionary<string, object> dictionary = new Dictionary<string, object>();
+
+            string key = "";
+            object value = null;
+
+            while (json.Read())
+            {
+                JsonTokenType tokenType = json.TokenType;
+                ReadOnlySpan<byte> valueSpan = json.Value;
+                switch (tokenType)
+                {
+                    case JsonTokenType.PropertyName:
+                        key = valueSpan.ConvertToString();
+                        dictionary.Add(key, null);
+                        break;
+                    case JsonTokenType.Value:
+                        JsonValueType valueType = json.ValueType;
+
+                        switch (valueType)
+                        {
+                            case JsonValueType.True:
+                            case JsonValueType.False:
+                                value = valueSpan.ConvertToBool();
+                                break;
+                            case JsonValueType.Number:
+                                NumberType type;
+                                (type, value) = valueSpan.GetNumberType();
+                                break;
+                            case JsonValueType.String:
+                                value = valueSpan.ConvertToString();
+                                break;
+                            case JsonValueType.Null:
+                                break;
+                        }
+
+                        if (dictionary.TryGetValue(key, out _))
+                        {
+                            dictionary[key] = value;
+                        }
+                        else
+                        {
+                            dictionary.Add(key, value);
+                        }
+
+                        break;
+                    case JsonTokenType.StartObject:
+                        value = JsonLabReaderDictionaryLoop(ref json);
+                        if (dictionary.TryGetValue(key, out _))
+                        {
+                            dictionary[key] = value;
+                        }
+                        else
+                        {
+                            dictionary.Add(key, value);
+                        }
+                        break;
+                    case JsonTokenType.StartArray:
+                        value = JsonLabReaderListLoop(ref json);
+                        if (dictionary.TryGetValue(key, out _))
+                        {
+                            dictionary[key] = value;
+                        }
+                        else
+                        {
+                            dictionary.Add(key, value);
+                        }
+                        break;
+                    case JsonTokenType.EndObject:
+                        return dictionary;
+                    case JsonTokenType.None:
+                    case JsonTokenType.Comment:
+                    default:
+                        break;
+                }
+            }
+            return dictionary;
+        }
+
+        public static List<object> JsonLabReaderListLoop(ref Utf8JsonReader json)
+        {
+            List<object> arrayList = new List<object>();
+
+            object value = null;
+
+            while (json.Read())
+            {
+                JsonTokenType tokenType = json.TokenType;
+                ReadOnlySpan<byte> valueSpan = json.Value;
+                switch (tokenType)
+                {
+                    case JsonTokenType.Value:
+                        JsonValueType valueType = json.ValueType;
+
+                        switch (valueType)
+                        {
+                            case JsonValueType.True:
+                            case JsonValueType.False:
+                                value = valueSpan.ConvertToBool();
+                                break;
+                            case JsonValueType.Number:
+                                NumberType type;
+                                (type, value) = valueSpan.GetNumberType();
+                                break;
+                            case JsonValueType.String:
+                                value = valueSpan.ConvertToString();
+                                break;
+                            case JsonValueType.Null:
+                                break;
+                        }
+
+                        arrayList.Add(value);
+
+                        break;
+                    case JsonTokenType.StartObject:
+                        value = JsonLabReaderDictionaryLoop(ref json);
+                        arrayList.Add(value);
+                        break;
+                    case JsonTokenType.StartArray:
+                        value = JsonLabReaderListLoop(ref json);
+                        arrayList.Add(value);
+                        break;
+                    case JsonTokenType.EndArray:
+                        return arrayList;
+                    case JsonTokenType.None:
+                    case JsonTokenType.Comment:
+                    default:
+                        break;
+                }
+            }
+            return arrayList;
+        }
+
+        public static byte[] JsonLabReturnBytesHelper(byte[] data, out int length)
+        {
+            var reader = new Utf8JsonReader(data);
+            return JsonLabReaderLoop(data.Length, out length, ref reader);
+        }
+
+        public static object JsonLabReturnObjectHelper(byte[] data)
+        {
+            var reader = new Utf8JsonReader(data);
+            return JsonLabReaderLoop(ref reader);
         }
     }
 }
