+++ tests/System.Text.JsonLab.Tests/TestCaseType.cs
@@ -1,0 +1,26 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Text.JsonLab.Tests
+{
+    public enum TestCaseType
+    {
+        HelloWorld,
+        Basic,
+        BasicLargeNum,
+        SpecialNumForm,
+        SpecialStrings,
+        ProjectLockJson,
+        FullSchema1,
+        FullSchema2,
+        DeepTree,
+        BroadTree,
+        LotsOfNumbers,
+        LotsOfStrings,
+        Json400B,
+        Json4KB,
+        Json40KB,
+        Json400KB,
+    }
+}
